Change key field of tuple-ref from u64 to u32
I don't think this has a significant impact on the implementation,
especially given that the tuple index doesn't actually get stored
anywhere within the WebAssembly (at least under the current
implementation), but I could see this changing in the future.
I am also trying to simplify the design, and it seems reasonable to
limit tuples to having at most 2^32 possible elements.

+++ src/common.rs
@@ -113,7 +113,7 @@ pub enum ExprKind<E: ExprMeta> {
     Null(Type),
     FnApp(E, Vector<E>),         // func, arguments
     Tuple(Vector<E>),            // list of expressions, type annotation
-    TupleGet(E, u64),            // env, index - index must explicitly be a number
+    TupleGet(E, u32),            // env, index - index must explicitly be a number
     Pack(E, Type, Type),         // exp, type substitution, existential type
     Unpack(String, E, u64, E),   // new var, package, type var, body
     Record(Vector<(String, E)>), // map from values to labels
